Move zone checking code from pdnsutil to the common checking code.
This is purely mechanical and there are no other changes in there.
